requirements.txt
python-dotenv==0.15.0
requests==2.32.3
terminaltables==3.1.10